feat: update tests for weakened effects
