+++ packages/server/test/potions/potionEffects/greenPotion.test.ts
@@ -306,63 +306,85 @@ describe('Try to consume green potion in various cases', () => {
   });
 });
 
-// BLACK POTION TEST
+// describe('Try to consume an unknown potion that is similar to green potion', () => {
+//   test('Test weakened green potion effect', () => {
+//     FantasyDate.initialDate();
 
-describe('Try to consume black potion in various cases', () => {
-  test('Spawn and kill monster with a black potion', () => {
-    FantasyDate.initialDate();
+//     const playerPosition: Coord = { x: 0, y: 0 };
+//     const enemyPosition: Coord = { x: 0, y: 1 };
+//     const potionLocation: Coord = { x: 1, y: 0 };
 
-    const playerPosition: Coord = { x: 0, y: 0 };
-    const potionLocation: Coord = { x: 1, y: 0 };
+//     // create a fight initiator (blob -> hunt)
+//     mobFactory.makeMob('blob', playerPosition, 'TestingID', 'TestAttacker');
+//     const testAttacker = Mob.getMob('TestingID');
+//     expect(testAttacker).not.toBeNull();
 
-    // create a fight initiator (blob -> hunt)
-    mobFactory.makeMob('player', playerPosition, 'TestingID', 'MonsterSpawner');
-    const testMob = Mob.getMob('TestingID');
-    expect(testMob).not.toBeNull();
+//     // create a enemy (player)
+//     mobFactory.makeMob('player', enemyPosition, 'TestEnemyID', 'TestEnemy');
+//     const testEnemy = Mob.getMob('TestEnemyID');
+//     expect(testEnemy).not.toBeNull();
 
-    // create a potion
-    itemGenerator.createItem({
-      type: 'potion',
-      subtype: String(hexStringToNumber('#166060')),
-      position: potionLocation,
-      carriedBy: testMob
-    });
-    const potion = Item.getItemIDAt(potionLocation);
-    expect(potion).not.toBeNull();
-    const potionItem = Item.getItem(potion!);
-    expect(potionItem).not.toBeNull();
+//     // make the blob fight the player (due to low favorability)
+//     Community.makeFavor('alchemists', 'blobs', -100);
 
-    // ensure the initiator is carrying the potion
-    expect(testMob!.carrying).not.toBeNull();
-    expect(testMob!.carrying!.type).toBe('potion');
-    expect(testMob!.carrying!.subtype).toBe(
-      String(hexStringToNumber('#166060'))
-    );
+//     // create a potion
+//     itemGenerator.createItem({
+//       type: 'potion',
+//       subtype: String(hexStringToNumber('#00ff22')),
+//       position: potionLocation,
+//       carriedBy: testAttacker
+//     });
+//     const potion = Item.getItemIDAt(potionLocation);
+//     expect(potion).not.toBeNull();
+//     const potionItem = Item.getItem(potion!);
+//     expect(potionItem).not.toBeNull();
 
-    // have the attacker drink the potion
-    const testDrink = new Drink();
-    const test = testDrink.interact(testMob!, potionItem!);
-    expect(test).toBe(true);
+//     // have the attacker drink the potion
+//     const testDrink = new Drink();
+//     const test = testDrink.interact(testAttacker!, potionItem!);
+//     expect(test).toBe(true);
 
-    // check to make sure potion is not being carried
-    expect(testMob!.carrying).toBeUndefined();
+//     // grab health of enemy, see if it has started yet or we need one more tick
+//     const initEnemyHealth = testEnemy!.health;
+//     expect(initEnemyHealth).toBe(testEnemy!._maxHealth);
 
-    // check that monster exists
-    const monster = Mob.getMob('Monster');
-    expect(monster).not.toBeNull();
+//     // go through ticks to ensure attacker is in range and attacking
+//     testAttacker?.tick(500);
+//     testEnemy?.tick(500);
+//     FantasyDate.runTick();
+//     expect(testAttacker!.action).toBe('hunt');
 
-    //wait to make the monster time out
-    // run ticks
-    for (let i = 0; i < 120; i++) {
-      FantasyDate.runTick();
-    }
-    monster?.tick(500);
+//     // check effects
+//     expect(testAttacker!.damageOverTime).toBe(1*0.5);
+//     expect(testEnemy!.poisoned).toBe(1);
 
-    // check to make sure monster is dead
-    const deadMonster = Mob.getMob('Monster');
-    expect(deadMonster?.action).toBe('destroyed');
-  });
-});
+//     // kill attacker so poison is the only thing doing damage
+//     testAttacker!.destroy();
+
+//     // grab health of enemy, make sure it has changed
+//     const firstEnemyHealth = testEnemy!.health;
+//     expect(firstEnemyHealth).toBeLessThan(testEnemy!._maxHealth);
+
+//     // run a tick
+//     testEnemy!.tick(500);
+//     FantasyDate.runTick();
+
+//     // health should have decreased again
+//     const secondEnemyHealth = testEnemy!.health;
+//     expect(secondEnemyHealth).toBeLessThan(firstEnemyHealth);
+
+//     // run a tick
+//     testEnemy!.tick(500);
+//     FantasyDate.runTick();
+
+//     // health should be the same (effect ran out)
+//     const thirdEnemyHealth = testEnemy!.health;
+//     expect(thirdEnemyHealth).toBe(secondEnemyHealth);
+
+//     // should have run out now
+//     expect(testEnemy!.poisoned).toBe(0);
+//   });
+// });
 
 afterAll(() => {
   DB.close();
