+++ packages/server/test/potions/potionEffects/orangePotion.test.ts
@@ -159,6 +159,58 @@ describe('Try to consume orange potion in various cases', () => {
   });
 });
 
+describe('Try to consume an undefined potion similar to an orange potion', () => {
+  test('Test weakened orange potion effect', () => {
+    FantasyDate.initialDate();
+    const position: Coord = { x: 0, y: 0 };
+    const potionLocation: Coord = { x: 1, y: 0 };
+
+    // create a player
+    mobFactory.makeMob('player', position, 'TestID', 'TestPlayer');
+    const testMob = Mob.getMob('TestID');
+    expect(testMob).not.toBeNull();
+
+    // create a potion
+    itemGenerator.createItem({
+      type: 'potion',
+      subtype: String(hexStringToNumber('#e79611')),
+      position: potionLocation,
+      carriedBy: testMob
+    });
+    const potion = Item.getItemIDAt(potionLocation);
+    expect(potion).not.toBeNull();
+    const potionItem = Item.getItem(potion!);
+    expect(potionItem).not.toBeNull();
+
+    // ensure the player is carrying the potion
+    expect(testMob!.carrying).not.toBeNull();
+    expect(testMob!.carrying!.type).toBe('potion');
+    expect(testMob!.carrying!.subtype).toBe(
+      String(hexStringToNumber('#e79611'))
+    );
+
+    // set initial attack
+    const startAttack = testMob!._attack; // should be 5 at default
+
+    // have the player drink the potion
+    const testDrink = new Drink();
+    const test = testDrink.interact(testMob!, potionItem!);
+    expect(test).toBe(true);
+
+    for (let i = 0; i < 15; i++) {
+      // 15 ticks to check stacking
+      FantasyDate.runTick();
+    }
+    testMob?.tick(500);
+
+    // check to make sure potion is not being carried
+    expect(testMob!.carrying).toBeUndefined();
+
+    // check attributes on player (should be boosted)
+    expect(testMob!._attack).toBe(startAttack + startAttack * 0.5 * 0.5);
+  });
+});
+
 afterAll(() => {
   DB.close();
 });
